Rename the tests according to the project name
